+++ src/OpenTelemetry/Metrics/HistogramBucket.cs
@@ -1,4 +1,4 @@
-// <copyright file="HistogramMeasurement.cs" company="OpenTelemetry Authors">
+// <copyright file="HistogramBucket.cs" company="OpenTelemetry Authors">
 // Copyright The OpenTelemetry Authors
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
@@ -16,9 +16,9 @@
 
 namespace OpenTelemetry.Metrics
 {
-    public readonly struct HistogramMeasurement
+    public readonly struct HistogramBucket
     {
-        internal HistogramMeasurement(double explicitBound, long bucketCount)
+        internal HistogramBucket(double explicitBound, long bucketCount)
         {
             this.ExplicitBound = explicitBound;
             this.BucketCount = bucketCount;