+++ src/OpenTelemetry/Metrics/HistogramBuckets.cs
@@ -1,4 +1,4 @@
-// <copyright file="HistogramMeasurements.cs" company="OpenTelemetry Authors">
+// <copyright file="HistogramBuckets.cs" company="OpenTelemetry Authors">
 // Copyright The OpenTelemetry Authors
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
@@ -16,7 +16,7 @@
 
 namespace OpenTelemetry.Metrics
 {
-    public class HistogramMeasurements
+    public class HistogramBuckets
     {
         internal readonly long[] BucketCounts;
 
@@ -34,7 +34,7 @@ namespace OpenTelemetry.Metrics
 
         internal double Sum;
 
-        internal HistogramMeasurements(double[] histogramBounds)
+        internal HistogramBuckets(double[] histogramBounds)
         {
             this.ExplicitBounds = histogramBounds;
             this.BucketCounts = histogramBounds != null ? new long[histogramBounds.Length + 1] : null;
@@ -46,33 +46,31 @@ namespace OpenTelemetry.Metrics
 
         public struct Enumerator
         {
-            private readonly bool isHistogramSumCount;
             private readonly int numberOfBuckets;
             private readonly int numberofExplicitBounds;
-            private readonly HistogramMeasurements histogramMeasurements;
+            private readonly HistogramBuckets histogramMeasurements;
             private int index;
 
-            internal Enumerator(HistogramMeasurements histogramMeasurements)
+            internal Enumerator(HistogramBuckets histogramMeasurements)
             {
                 this.histogramMeasurements = histogramMeasurements;
                 this.index = 0;
                 this.Current = default;
-                this.isHistogramSumCount = histogramMeasurements.ExplicitBounds == null;
-                this.numberOfBuckets = this.isHistogramSumCount ? default : histogramMeasurements.BucketCounts.Length;
-                this.numberofExplicitBounds = this.isHistogramSumCount ? default : histogramMeasurements.ExplicitBounds.Length;
+                this.numberOfBuckets = histogramMeasurements.ExplicitBounds == null ? 0 : histogramMeasurements.BucketCounts.Length;
+                this.numberofExplicitBounds = histogramMeasurements.ExplicitBounds == null ? 0 : histogramMeasurements.ExplicitBounds.Length;
             }
 
-            public HistogramMeasurement Current { get; private set; }
+            public HistogramBucket Current { get; private set; }
 
             public bool MoveNext()
             {
-                if (!this.isHistogramSumCount && this.index < this.numberOfBuckets)
+                if (this.index < this.numberOfBuckets)
                 {
                     double explicitBound = this.index < this.numberofExplicitBounds
                         ? this.histogramMeasurements.ExplicitBounds[this.index]
                         : double.PositiveInfinity;
                     long bucketCount = this.histogramMeasurements.AggregatedBucketCounts[this.index];
-                    this.Current = new HistogramMeasurement(explicitBound, bucketCount);
+                    this.Current = new HistogramBucket(explicitBound, bucketCount);
                     this.index++;
                     return true;
                 }