+++ src/OpenTelemetry/CHANGELOG.md
@@ -14,9 +14,9 @@
   aggregation is not implemented yet.
   ([#2660](https://github.com/open-telemetry/opentelemetry-dotnet/pull/2660))
 
-* Refactored `HistogramMeasurements` to provide an enumerator for enumerating
-  the BucketCounts and ExplicitBounds. Removed `GetBucketCounts` and
-  `GetExplicitBounds` methods from `MetricPoint`.
+* Renamed `HistogramMeasurements` to `HistogramBuckets` and added an enumerator
+  of type `HistogramBucket` for enumerating `BucketCounts` and `ExplicitBounds`.
+  Removed `GetBucketCounts` and `GetExplicitBounds` methods from `MetricPoint`.
   ([#2664](https://github.com/open-telemetry/opentelemetry-dotnet/pull/2664))
 
 ## 1.2.0-beta2